fix(rdpgfx): Progressive decoder parity with FreeRDP — 5th mosaic root cause
Root cause (single-variable A/B on clean commit 3e90c94, real Windows session):
WTS2 Progressive (codecId 0x0009) emitted 64-pixel-aligned gray bands that
overwrote correctly decoded ClearCodec tiles. Gray coverage: 94.6% with
Progressive enabled, 0% with Progressive skipped. 8 live payloads captured.

Progressive decoder fixes (rfx_progressive.go, parity with FreeRDP
libfreerdp/codec/progressive.c):

- progressive quant table now parsed (16 bytes per entry, 3 component quants)
- quant combination: quant + progQuant - 1 (FreeRDP rule)
- extrapolate subband layout: 1023/1023/961/272/272/256/72/72/64/81
  (was non-extrapolate 1024/1024/1024/256/256/256/64/64/64/64)
- extrapolate DWT: rfxInverseDWT2DExtrapolate
- differential decode: separated from dequant (FreeRDP order)
- tile cache: sign/raw coefficients per tile for TILE_UPGRADE re-render
- TILE_UPGRADE: partial — re-renders from cached raw with updated quant;
  full SRL bit-plane upgrade state machine is not yet ported (upgrade
  frames are ~50% of captured stream)
- WTS2 payload capture: bounded ring buffer, JS export for offline replay
- offline replay test: progressive_replay_test.go (PROGRESSIVE_REPLAY_DIR)

Also:
- render_bridge: pass WASM pointer as float64 (go/wasm int is 32-bit,
  values >= 2^31 cast to negative JS int, producing wrong memory views)
- RDP_STATUS.md: complete 5-layer root cause documentation
- Progressive replay test + demo harness

Known limitation: TILE_UPGRADE SRL state machine not yet ported.
Production should block Progressive codec capability until full parity
is achieved, or the server will still send upgrade frames.

+++ rdp-wasm/render_bridge.go
@@ -67,10 +67,15 @@ func forwardRenderEvent(event rdpgfx.RenderEvent) {
 	}
 	slog.Info("forwardRenderEvent", "kind", event.Kind, "frameId", event.FrameID, "surfaceId", event.SurfaceID, "dataLen", len(event.Data))
 	if (event.Kind == rdpgfx.RenderBitmap || event.Kind == rdpgfx.RenderClassicBitmap) && len(event.Data) > 0 && bitmapCallback.Type() == js.TypeFunction {
-		pointer := uintptr(unsafe.Pointer(unsafe.SliceData(event.Data)))
+		// Pass the linear-memory offset as float64, NOT int. On go/wasm,
+		// int is 32-bit; casting a uintptr ≥ 2^31 yields a negative int,
+		// and JS then builds Uint8Array(buffer, negativeOffset) → wrong
+		// bytes / empty views → full-screen garbage tiles. float64 can
+		// represent every wasm32 address exactly.
+		pointer := float64(uintptr(unsafe.Pointer(unsafe.SliceData(event.Data))))
 		bitmapCallback.Invoke(js.ValueOf(map[string]any{
 			"kind": int(event.Kind), "frameId": int(event.FrameID), "surfaceId": int(event.SurfaceID),
-			"rect": renderRectValue(event.Rect), "pointer": int(pointer),
+			"rect": renderRectValue(event.Rect), "pointer": pointer,
 			"length": len(event.Data), "stride": int(event.Stride),
 		}))
 		runtime.KeepAlive(event.Data)
